# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Road Hazard Monitoring System API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type DetectionSeverity = typeof DetectionSeverity[keyof typeof DetectionSeverity];


export const DetectionSeverity = {
  Low: 'Low',
  Medium: 'Medium',
  High: 'High',
  Critical: 'Critical',
} as const;

export interface Detection {
  id: number;
  road_name: string;
  latitude: number;
  longitude: number;
  pothole_count: number;
  avg_diameter_cm: number;
  severity: DetectionSeverity;
  severity_score: number;
  detected_at: string;
  /** @nullable */
  original_image_url?: string | null;
  /** @nullable */
  detected_image_url?: string | null;
  /** @nullable */
  mask_image_url?: string | null;
}

export type DetectionInputSeverity = typeof DetectionInputSeverity[keyof typeof DetectionInputSeverity];


export const DetectionInputSeverity = {
  Low: 'Low',
  Medium: 'Medium',
  High: 'High',
  Critical: 'Critical',
} as const;

export interface DetectionInput {
  road_name: string;
  latitude: number;
  longitude: number;
  pothole_count: number;
  avg_diameter_cm: number;
  severity: DetectionInputSeverity;
  severity_score?: number;
  /** @nullable */
  original_image_url?: string | null;
  /** @nullable */
  detected_image_url?: string | null;
  /** @nullable */
  mask_image_url?: string | null;
}

export type PotholeSeverity = typeof PotholeSeverity[keyof typeof PotholeSeverity];


export const PotholeSeverity = {
  Low: 'Low',
  Medium: 'Medium',
  High: 'High',
  Critical: 'Critical',
} as const;

export interface Pothole {
  id: number;
  detection_id: number;
  pothole_code: string;
  diameter_cm: number;
  area_m2: number;
  perimeter_m: number;
  mask_coverage: number;
  confidence: number;
  severity: PotholeSeverity;
}

export interface RoadScore {
  rank: number;
  road_name: string;
  severity_score: number;
  pothole_count: number;
  dominant_severity: string;
}

export interface RoadRanking {
  worst: RoadScore[];
  best: RoadScore[];
}

export interface FeedItem {
  id: number;
  road_name: string;
  pothole_count: number;
  severity: string;
  detected_at: string;
  latitude: number;
  longitude: number;
}

export type DashboardSummarySeverityBreakdown = {
  Low: number;
  Medium: number;
  High: number;
  Critical: number;
};

export interface DashboardSummary {
  total_roads: number;
  total_potholes: number;
  critical_zones: number;
  avg_diameter_cm: number;
  /** @nullable */
  last_detection_at: string | null;
  severity_breakdown: DashboardSummarySeverityBreakdown;
}

export type ListPotholesParams = {
/**
 * @nullable
 */
detection_id?: number | null;
};

export type GetDetectionFeedParams = {
/**
 * @nullable
 */
limit?: number | null;
};