# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Road Hazard Monitoring System API
 * OpenAPI spec version: 0.1.0
 */
import type { DashboardSummarySeverityBreakdown } from './dashboardSummarySeverityBreakdown';

export interface DashboardSummary {
  total_roads: number;
  total_potholes: number;
  critical_zones: number;
  avg_diameter_cm: number;
  /** @nullable */
  last_detection_at: string | null;
  severity_breakdown: DashboardSummarySeverityBreakdown;
}