# lib/api-zod/src/generated/types/detectionInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Road Hazard Monitoring System API
 * OpenAPI spec version: 0.1.0
 */
import type { DetectionInputSeverity } from './detectionInputSeverity';

export interface DetectionInput {
  road_name: string;
  latitude: number;
  longitude: number;
  pothole_count: number;
  avg_diameter_cm: number;
  severity: DetectionInputSeverity;
  severity_score?: number;
  /** @nullable */
  original_image_url?: string | null;
  /** @nullable */
  detected_image_url?: string | null;
  /** @nullable */
  mask_image_url?: string | null;
}