# lib/api-zod/src/generated/types/feedItem.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Road Hazard Monitoring System API
 * OpenAPI spec version: 0.1.0
 */

export interface FeedItem {
  id: number;
  road_name: string;
  pothole_count: number;
  severity: string;
  detected_at: string;
  latitude: number;
  longitude: number;
}