# lib/api-zod/src/generated/types/pothole.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Road Hazard Monitoring System API
 * OpenAPI spec version: 0.1.0
 */
import type { PotholeSeverity } from './potholeSeverity';

export interface Pothole {
  id: number;
  detection_id: number;
  pothole_code: string;
  diameter_cm: number;
  area_m2: number;
  perimeter_m: number;
  mask_coverage: number;
  confidence: number;
  severity: PotholeSeverity;
}