# lib/api-zod/src/generated/types/roadRanking.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Road Hazard Monitoring System API
 * OpenAPI spec version: 0.1.0
 */
import type { RoadScore } from './roadScore';

export interface RoadRanking {
  worst: RoadScore[];
  best: RoadScore[];
}