# lib/api-zod/src/generated/types/roadScore.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Road Hazard Monitoring System API
 * OpenAPI spec version: 0.1.0
 */

export interface RoadScore {
  rank: number;
  road_name: string;
  severity_score: number;
  pothole_count: number;
  dominant_severity: string;
}